IOS-3451: Move protocol into separate file
